Removes environment object and custom log emitter
- Simplified environment setting logic to have everything on except
optimize memory on SharedEnv.Default. This will me that is a user does
not provide a setting for that environment variable we will use our
default. This does leave something to be desired for the in the logging
but I think that I would that an update in the log emitter makes more
sense than complicated logic in the buildpack.
- also removes usage of ioutil